feat: v1 transaction response structs

+++ client/one/transactions_requests.go
@@ -15,16 +15,16 @@ type GetUnconfirmedTransactionsQuery struct {
 
 type GetTransactionsQuery struct {
 	BlockId         string `url:"blockId"`
-	Limit           int    `url:"limit"`
-	Type            int    `url:"type"`
+	Limit           byte   `url:"limit"`
+	Type            byte   `url:"type"`
 	OrderBy         string `url:"orderBy"`
-	Offset          int    `url:"offset"`
+	Offset          byte   `url:"offset"`
 	SenderPublicKey string `url:"senderPublicKey"`
 	VendorField     string `url:"vendorField"`
 	OwnerPublicKey  string `url:"ownerPublicKey"`
 	OwnerAddress    string `url:"ownerAddress"`
 	SenderId        string `url:"senderId"`
 	RecipientId     string `url:"recipientId"`
-	Amount          int    `url:"amount"`
-	Fee             int    `url:"fee"`
+	Amount          byte   `url:"amount"`
+	Fee             byte   `url:"fee"`
 }
